Use an arbitrarily-chosen value for the block size of arrays.
Use an arbitrarily-chosen value for the block size of arrays in memory,
which don't have an inherent block size in the way that filesystems do.
In the future this value may be tuned for performance, but until then,
an arbitrarily chosen value is no less arbitrary than the page size,
especially since I/O arrays are not page-aligned.

This avoids a dependency on the `page_size` crate.